requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Cors>=4.0.0
mysql-connector-python>=8.2.0
openai>=1.12.0
python-dotenv>=1.0.1
redis>=5.0.1
rq>=1.16.1
waitress>=3.0.0
gunicorn 
gevent
python-docx